+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * All `AuthSrc()` derivatives, deprecated and never used, have been removed.
 * New module `mat` adds a shallow reimplementation of `Matrix()` in order to implement matrix multiplication
 * Removed obsolete `configparse` implementation that has been around since 0.3 and shelved since 0.4.
+* `color`: added support for conversion from RGB to sRGB
 
 ## 0.11.2